Fix hellbanned user able to request with short pending queue
When queueing an unlocked song and the number of songs in the pending
queue is shorter than MIN_QUEUE_SONGS_LIMIT the song is forced on the
queue. Forcing songs on the queue skips the check for the hellban status
of a user. Fix by checking the hellban status earlier.