Only attempt to set song data again if a file is required
A song file's metadata is read again after editing a song through the
admin site but only if status is not Kaput. This results in trying to
read metadata for sites that still only have the legacy_flag set to
Missing (and status set to Active instead of Kaput). Fix by making use
of song.requires_file().