Improvements for non-Kaput/Missing songs without file
Currently whether or not to re-read or zero-out metadata is also based
on the status of the song. While this works now it's not correct when
later on songs with any non-playable status plus file not set are
accepted, as then the file's object (set to None) gets accessed.
Fix by simply checking if file is set, as a pre-condition before
checking if file is an existing file.