Update legacy_flag automatically on admin site
Legacy flag should only be changed to N/A via the admin site. Setting
a song to Missing (not having a file) is not possible and setting it to
Recovered should be done through other means. To prevent editing
mistakes from being made make the legacy_flag field read-only and
automatically update legacy_flag to N/A if a replacement file is
provided.