Always generate prelisten file if one is desired
In case create_in_background() is never called a prelisten file is also
never generated because templates only check for existence of the file.
Fix by combining create_in_background() and exists() into generate()
such that a prelisten file is still only generated if needed and the
function returns True if it already exists, False otherwise.