Fix prelistening to songs with a non-existing file
If a song has an MP3 file set that doesn't exist a symlink gets created
that points to a non-existing file. Upon re-entry in generate() the
prelisten file then still doesn't exist while the symlink does. This
results in a File Exists error when trying to create the symlink again.
Fix by making the Prelisten object invalid if the song file doesn't
exist.